% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upmdata.R
\docType{data}
\name{upmdata}
\alias{upmdata}
\title{The dataset `upmdata.Rdata` contains example data from a theoretical tuberculosis household
contact study of 800 households contacts of an infectious pulmonary TB case from 108 households.
The first four variables are numeric. The final predictor varialbe is binary.
Next is the data defining household: cluster, and finally the measured outcome
of a Tuberculin Skin Test (TST/PPD) for each contact, y. In these data, the true proportion
of contacts contracting TB from outside the home is 15.5 percent. The true coeffient values are, in order
b0 = -1.5, b1 =  0.15, b2 = -2.25, b3 = 0 (no effect), b4 =  0.2, b5 = -0.1.}
\format{An object of class \code{data.frame} with 800 rows and 7 columns.}
\usage{
upmdata
}
\description{
See the introductory vignette for this package for more information.
}
\keyword{datasets}
